Cleaned up shell completion
Since we use "kubernetes" as a selection which installs kubectl and
kubeadm, we need to include those for generating the shell completion
info.

NOTE: These tasks will FAIL if kubernetes is not selected, since it
tries to run "kubeadm completion" just because kubeadm is in the list.
Need to add "ignore_errors: true" to the tasks, which is a bit dirty ...