Now using a ROLE_PATH env var to point to role location
The role/playbook being tested is mounted at /drone/<rolename> by drone
so the main working dir is /drone.  The role is included cleanly by molecule
via import_role using the role name which works fine.  But importing a playbook
requires import_playbook with a path to the exact playbook file.  As in
/<working dir>/path/path/playbook.yml which in this case equated to
/drone/cloud_tools.yml

This ROLE_PATH var just points directly to the playbook.