Switching from docker to podman for molecule testing
We still use the local docker setup to spin up our testing env using the
halfwalker/alpineset2 image - new version of toolset2.  But when
molecule creates containers it uses podman and creates them right in the
alpineset2 container - no more reaching out to the host docker or
needing a docker-in-docker sidecar container.